+++ src/store/auth/authSlice.js
@@ -1,0 +1,27 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+export const authSlice = createSlice({
+   name: 'auth',
+   initialState: {// para rvisar si el user esta auth
+       status: 'checking', //'not-authenticated', 'authenticated'
+       uid: null,
+       email: null,
+       displayName: null,
+       photoURL: null,
+       errorMessage: null,
+   },
+   reducers: {
+      login: ( state, action ) =>{
+
+      },
+      logout: ( state, payload ) => {
+
+      },
+      checkingCredentials: ( state ) => {
+
+      }
+   }
+});
+
+
+export const { login, logout, checkingCredentials } = authSlice.actions;