+++ src/store/store.js
@@ -1,0 +1,10 @@
+import { configureStore } from '@reduxjs/toolkit'
+import { authSlice } from './auth'
+
+// import rootReducer from './reducers'
+
+export const store = configureStore({
+  reducer: {
+    auth: authSlice.reducer,
+  },
+})