+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 - Se agrega el proyecto inicial MaterialUI, estructura y diseño
+
+- Se instala redux al proyecto:
+    yarn add @reduxjs/toolkit
+    yarn add react-redux
+    Creacion del Store y Slice de auth
+    Add src/store/store.js
+    Add src/store/auth/authSlice.js